Feat(Descriptor API): Create descriptor-api sub-project and move BlackDuckProviderKey there
