+++ alert-database/build.gradle
@@ -21,6 +21,7 @@ apply plugin: com.bmuschko.gradle.docker.DockerRemoteApiPlugin
 
 dependencies {
     implementation platform(project(':alert-platform'))
+    implementation project(':descriptor-api')
 
     implementation 'com.synopsys.integration:blackduck-common', rootProject.ext.blackduckCommonExcludes
     implementation 'org.springframework.data:spring-data-jpa'
