+++ descriptor-api/build.gradle
@@ -1,0 +1,8 @@
+ext.moduleName = 'com.synopsys.integration.alert.descriptor.api'
+
+dependencies {
+    implementation platform(project(':alert-platform'))
+    implementation project(':alert-common')
+    implementation 'org.springframework:spring-context'
+    implementation 'com.synopsys.integration:blackduck-common', rootProject.ext.blackduckCommonExcludes
+}
