+++ provider/build.gradle
@@ -5,6 +5,7 @@ dependencies {
 
     implementation project(':alert-common')
     implementation project(':alert-database')
+    implementation project(':descriptor-api')
 
     implementation 'org.springframework:spring-context'
     implementation 'org.springframework:spring-web'
