+++ provider/src/main/java/com/synopsys/integration/alert/provider/blackduck/action/BlackDuckConfigurationAction.java
@@ -26,7 +26,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import com.synopsys.integration.alert.common.action.ConfigurationAction;
-import com.synopsys.integration.alert.provider.blackduck.BlackDuckProviderKey;
+import com.synopsys.integration.alert.descriptor.api.BlackDuckProviderKey;
 
 @Component
 public class BlackDuckConfigurationAction extends ConfigurationAction {
