+++ provider/src/main/java/com/synopsys/integration/alert/provider/blackduck/descriptor/BlackDuckDescriptor.java
@@ -26,7 +26,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import com.synopsys.integration.alert.common.descriptor.ProviderDescriptor;
-import com.synopsys.integration.alert.provider.blackduck.BlackDuckProviderKey;
+import com.synopsys.integration.alert.descriptor.api.BlackDuckProviderKey;
 
 @Component
 public class BlackDuckDescriptor extends ProviderDescriptor {
