+++ src/test/java/com/synopsys/integration/alert/mock/MockConfigurationModelFactory.java
@@ -26,7 +26,7 @@ import com.synopsys.integration.alert.common.persistence.model.ConfigurationJobM
 import com.synopsys.integration.alert.common.persistence.model.ConfigurationModel;
 import com.synopsys.integration.alert.common.persistence.model.mutable.ConfigurationModelMutable;
 import com.synopsys.integration.alert.common.util.DataStructureUtils;
-import com.synopsys.integration.alert.provider.blackduck.BlackDuckProviderKey;
+import com.synopsys.integration.alert.descriptor.api.BlackDuckProviderKey;
 import com.synopsys.integration.blackduck.api.manual.enumeration.NotificationType;
 
 public class MockConfigurationModelFactory {
@@ -227,4 +227,5 @@ public class MockConfigurationModelFactory {
         configurationFieldModel.setFieldValue(fieldValue);
         return configurationFieldModel;
     }
+
 }
