+++ src/test/java/com/synopsys/integration/alert/performance/utility/BlackDuckProviderService.java
@@ -14,7 +14,7 @@ import com.synopsys.integration.alert.common.enumeration.ConfigContextEnum;
 import com.synopsys.integration.alert.common.rest.model.FieldModel;
 import com.synopsys.integration.alert.common.rest.model.FieldValueModel;
 import com.synopsys.integration.alert.common.rest.model.MultiFieldModel;
-import com.synopsys.integration.alert.provider.blackduck.BlackDuckProviderKey;
+import com.synopsys.integration.alert.descriptor.api.BlackDuckProviderKey;
 import com.synopsys.integration.alert.util.TestProperties;
 import com.synopsys.integration.alert.util.TestPropertyKey;
 import com.synopsys.integration.bdio.model.Forge;
@@ -177,4 +177,5 @@ public class BlackDuckProviderService {
     public String getBlackDuckProjectVersion() {
         return blackDuckProjectVersion;
     }
+
 }
