+++ workflow/build.gradle
@@ -1,4 +1,4 @@
-ext.moduleName = 'com.synopsys.integration.alert.component'
+ext.moduleName = 'com.synopsys.integration.alert.workflow'
 
 dependencies {
     implementation platform(project(':alert-platform'))
